+++ src/pages/OrderPage.tsx
@@ -9,8 +9,8 @@ import { spacing } from '@/styles/spacing';
 import { typography } from '@/styles/typography';
 import GlobalStyle from '@/styles/GlobalStyle';
 import { useAuth } from '@/contexts/AuthContext';
-import { useProductSummaryQuery } from '@/api/productQuery';
-import { useCreateOrderMutation } from '@/api/orderQuery';
+import { useProductSummaryQuery } from '@/hooks/useProductQuery';
+import { useCreateOrderMutation } from '@/hooks/useOrderQuery';
 import type { OrderRequest } from '@/types/order';
 import orderCardsData from '@/data/orderCard';
 import ReceiverModal from './ReceiverModal';
